file_storage/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
python-multipart==0.0.20
